Move repository interfaces to TimeScout.Domain project
