--- TimeScout.API/Repository/IUserRepository.cs
@@ -1,17 +1,0 @@
-using TimeScout.API.Models;
-
-namespace TimeScout.API.Repository;
-
-public interface IUserRepository
-{
-    Task<bool> CheckIfUserExistsAsync(string email);
-    Task<User> CreateUserAsync(User newUser);
-    Task<bool> DeleteRefreshTokenAsync(int userId);
-    Task<bool> DeleteUserAsync(int userId);
-    Task<User> GetUserByEmailAndPasswordAsync(string email, string password);
-    Task<User> GetUserByIdAsync(int userId);
-    Task<User> GetUserByRefreshTokenAsync(string refreshToken);
-    Task<bool> RecoverUserAsync(int userId);
-    Task<bool> UpdateRefreshTokenAsync(int userId, string newRefreshToken);
-    Task<User> UpdateUserAsync(User user);
-}
