+++ TimeScout.Domain/Interfaces/IEventRepository.cs
@@ -1,6 +1,6 @@
-using TimeScout.API.Models;
+using TimeScout.Domain.Entities;
 
-namespace TimeScout.API.Repository;
+namespace TimeScout.Domain.Interfaces;
 
 public interface IEventRepository : IRepository<Event>
 {
